Address PR54 demo concurrency and client handling
